+++ sdk/cosmos/azure-cosmos/azure/cosmos/__init__.py
@@ -20,7 +20,7 @@
 # SOFTWARE.
 
 from ._version import VERSION
-from ._cosmos_responses import CosmosDictResponse, CosmosListResponse
+from ._cosmos_responses import CosmosDict, CosmosList
 from ._retry_utility import ConnectionRetryPolicy
 from .container import ContainerProxy
 from .cosmos_client import CosmosClient
@@ -65,7 +65,7 @@ __all__ = (
     "TriggerType",
     "ConnectionRetryPolicy",
     "ThroughputProperties",
-    "CosmosDictResponse",
-    "CosmosListResponse"
+    "CosmosDict",
+    "CosmosList"
 )
 __version__ = VERSION