+++ sdk/cosmos/azure-cosmos/azure/cosmos/_execution_context/base_execution_context.py
@@ -26,7 +26,7 @@ database service.
 from collections import deque
 import copy
 from .. import _retry_utility, http_constants
-from .._cosmos_responses import CosmosListResponse
+from .._cosmos_responses import CosmosList
 
 # pylint: disable=protected-access
 
@@ -77,10 +77,10 @@ class _QueryExecutionContextBase(object):
         QueryIterable has exposed fetch_next_block api).
 
         :return: List of results.
-        :rtype: ~azure.cosmos.CosmosListResponse
+        :rtype: ~azure.cosmos.CosmosList
         """
         self._ensure()
-        res = CosmosListResponse(self._buffer, response_headers=self._response_headers)
+        res = CosmosList(self._buffer, response_headers=self._response_headers)
         self._buffer.clear()
         return res
 
@@ -142,7 +142,7 @@ class _QueryExecutionContextBase(object):
                 self._continuation = None
             if fetched_items:
                 break
-        return CosmosListResponse(fetched_items, response_headers=response_headers)
+        return CosmosList(fetched_items, response_headers=response_headers)
 
     def _fetch_items_helper_with_retries(self, fetch_function):
         def callback():